openssl: added many config options and help text
The Config.in file was reworked.
Added help text to most of the options.
Added option to build most of the non-broken algorithms.
To avoid leaving too many options to the non-expert user, the
not-so-common options are hidden util 'choose less common build options'
gets ticked.

Some unused code, such as GOST engine stubs, were removed from the
default build, so the its package size dropped by 4.5K (on
mipsel_24kc).

The openwrt configurations patch was redone, to match the CONFIG_ARCH
string.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>