openssl: Enable selection of version to build
Now you can choose to build versions:
  1.1.0 (default) - stable
  1.0.2 - LTS - long-term support (legacy)
  1.1.1 - development (beta)

There are some changes to the way v. 1.0.2 gets built compared to the
previous, 1.0.2-only, package:
 * The fuzz directory is skipped during build.
 * Build with 'no-sha0', as it is deprecated, and should not be used.
 * Stripped build dirs from CFLAGS shown in 'openssl version' command.
 * Added specific target options for mips64, powerpc, i386
 * Removed SSL3 from build options
 * Added option to build engines, including the GOST engine.
 * Added options to choose to build less used ciphersuites (they were
   always disabled before)
 * Added the option to turn off assembly optimization, to reduce package
   size at the expense of speed.

For the 1.1.x series, it was added a patch to enable the option to
change the default ciphersuite list ordering to prefer ChaCha20 over
AES-GCM; it's off by default even though it should be on for 32-bit
platforms, because Chacha is 3x faster than AES-256 with an
equivalent strength.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>