openssl: more build/config options tinkering
Revert removing NPN support.

Moved CMS from other options to top as lua-openssl depends on it.

Changed SSE2 option from "disable" to "enable" for consistency.

Made TLS 1.3 depend on EC, as it will not be built otherwise.

Added an option to change the default ciphersuite list ordering to
prefer ChaCha20 over AES-GCM; it's off by default even though it should
be on for 32-bit platforms because it's 3x faster with an equivalent
strength.

Stripped CFLAGS from binary for 1.0.2, and changed the patch for 1.1.0
to be consistent with the other versions.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>