mwlwifi: Update the 88W8964's firmware to 9.3.2.12
Updates the 88W8964 firmware used in the Linksys WRT3200ACM and WRT32X
[v9.3.2.6 -> v9.3.2.12]

Removes 0c43219 ("mwlwifi: Fix loading with backports v5.3")
as it has been merged upstream.

Unfortunately, there is a bug wherein Kaloz's repo, the version
detection mechanism for fixing vendor commands doesn't work.

It pulls in the Linux kernel version, which as of this time is
"4.14.y" or "4.19.y"

However, the proper behaviour is that it should pull in the mac80211
backports version which as of now is "5.4-rc8"

The included patch works around this by removing the version check.

Signed-off-by: Jose Olivera <oliverajeo@gmail.com>
[ added a space to force failure when this is updated by openwrt ]
Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>