build: define RTC_SUPPORT as a bool
Currently, RTC_SUPPORT is defined as a tristate, with 'depends on m',
which is supposed to only let it be set to 'm' or 'n'.  However,
scripts/target-metadata.pl will 'select' it, or setting it to 'y', which
defeats it's 'depends on m' restriction.  The "users" of the symbol are
not expecting it to be necessarily 'm' either, so we can safely use it
as bool.  Current Linux 'conf' will issue a warning when it detects such
unmet dependencies, so this avoids that warning when updating our build
system.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>