build: make our config Makefile closer to upstream
This is done in preparation for an update of scripts/config.

scripts/config/Makefile has been patched in different ways so many times
that it is hard to keep in sync with changes upstream.  This attempts to
bring it, and other ancillary files, as close to upstream as possible.

It includes a change in toplevel.mk to find pkg-config for make xconfig
using the same strategy as the other config targets, i.e. by prepending
the PATH variable with the location of the distro's pkg-config instead
of passing it as a variable.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>