build: config - make build files close to upstream
Minimize the changes we're making to the upstream Makefile, and other
build files in scripts/config.

When doing 'make xconfig', instead of using a make variable to point to
the distro pkg-config, prepend its location to the PATH, and allow use
of plain 'pkg-config', just like it is done upstream, and here for
menuconfig, kernel_menuconfig, and kernel_nconfig.

The README file was updated to list the changes made from upstream.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>